progress card and be sample node
